pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = []
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
amqp = "^5.2.0"
annotated-types = "^0.6.0"
anyascii = "^0.3.2"
asgiref = "^3.7.2"
attrs = "^23.2.0"
autobahn = "^23.6.2"
Automat = "^22.10.0"
Babel = "^2.14.0"
beautifulsoup4 = "^4.11.2"
billiard = "^4.2.0"
celery = "^5.3.6"
certifi = "^2023.11.17"
cffi = "^1.16.0"
cfgv = "^3.4.0"
channels = "^4.0.0"
channels-redis = "^4.1.0"
charset-normalizer = "^3.3.2"
click = "^8.1.7"
click-didyoumean = "^0.3.0"
click-plugins = "^1.1.1"
click-repl = "^0.3.0"
constantly = "^23.10.4"
coreschema = "^0.0.4"
cryptography = "^41.0.7"
daphne = "^4.0.0"
defusedxml = "^0.7.1"
diff-match-patch = "^20230430"
distlib = "^0.3.8"
django-celery-beat = "^2.1.0"
django-colorfield = "^0.11.0"
django-cors-headers = "^3.14.0"
django-debug-toolbar = "^4.2.0"
django-extensions = "^3.2.3"
django-filter = "^23.5"
django-guardian = "^2.4.0"
django-import-export = "^3.3.4"
django-modelcluster = "^6.1"
django-money = "^3.4.1"
django-permissionedforms = "^0.1"
django-phonenumber-field = "^7.1.0"
django-simple-history = "^3.4.0"
django-taggit = "^4.0.0"
django-templated-mail = "^1.1.1"
django-timezone-field = "^4.2.3"
django-treebeard = "^4.7"
django-unfold = "^0.19.0"
djangoql = "^0.18.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
djoser = "^2.2.0"
dotty-dict = "^1.3.1"
draftjs-exporter = "^2.1.7"
drf-spectacular = "^0.27.0"
drf-yasg = "^1.21.6"
et-xmlfile = "^1.1.0"
filelock = "^3.13.1"
filetype = "^1.2.0"
gitdb = "^4.0.11"
GitPython = "^3.1.41"
gunicorn = "^21.2.0"
html5lib = "^1.1"
hyperlink = "^21.0.0"
identify = "^2.5.33"
idna = "^3.6"
importlib-resources = "^6.1.1"
incremental = "^22.10.0"
inflection = "^0.5.1"
Jinja2 = "^3.1.2"
jsonschema = "^4.20.0"
jsonschema-specifications = "^2023.12.1"
kombu = "^5.3.4"
l18n = "^2021.3"
markdown-it-py = "^3.0.0"
MarkupPy = "^1.14"
MarkupSafe = "^2.1.3"
mdurl = "^0.1.2"
msgpack = "^1.0.7"
nodeenv = "^1.8.0"
oauthlib = "^3.2.2"
odfpy = "^1.4.1"
openpyxl = "^3.1.2"
packaging = "^23.2"
phonenumbers = "^8.13.16"
pillow = "^10.2.0"
pillow-heif = "^0.14.0"
platformdirs = "^4.1.0"
ply = "^3.11"
pre-commit = "^3.6.0"
prompt-toolkit = "^3.0.43"
psycopg = "^3.1.17"
psycopg2 = "^2.9.9"
py-moneyed = "^3.0"
pyasn1 = "^0.5.1"
pyasn1-modules = "^0.3.0"
pycparser = "^2.21"
pydantic = "^2.5.3"
pydantic_core = "^2.14.6"
Pygments = "^2.17.2"
PyJWT = "^2.8.0"
pyOpenSSL = "^23.3.0"
python-crontab = "^3.0.0"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.0"
python-gitlab = "^4.3.0"
python-semantic-release = "^8.7.0"
python3-openid = "^3.2.0"
pytz = "^2023.3.post1"
PyYAML = "^6.0.1"
redis = "^5.0.1"
referencing = "^0.32.1"
requests = "^2.31.0"
requests-oauthlib = "^1.3.1"
requests-toolbelt = "^1.0.0"
rich = "^13.7.0"
rpds-py = "^0.16.2"
sentry-sdk = "^1.39.2"
service-identity = "^23.1.0"
shellingham = "^1.5.4"
six = "^1.16.0"
smmap = "^5.0.1"
social-auth-app-django = "^5.4.0"
social-auth-core = "^4.5.1"
soupsieve = "^2.5"
sqlparse = "^0.4.4"
tablib = "^3.5.0"
telepath = "^0.3.1"
tomlkit = "^0.12.3"
Twisted = "^23.10.0"
txaio = "^23.1.1"
typing_extensions = "^4.9.0"
tzdata = "^2023.4"
uritemplate = "^4.1.1"
urllib3 = "^2.1.0"
vine = "^5.1.0"
virtualenv = "^20.25.0"
wagtail = "^5.2.2"
wcwidth = "^0.2.13"
webencodings = "^0.5.1"
whitenoise = "^6.5.0"
Willow = "^1.6.3"
xlrd = "^2.0.1"
xlwt = "^1.3.0"

[tool.ruff]
fix = true
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    "**/migrations"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
